+++ aspecd/__init__.py
@@ -41,8 +41,3 @@ Utilities
     Exceptions for the ASpecD package
 
 """
-import logging
-
-
-package_logger = logging.getLogger(__name__)
-package_logger.addHandler(logging.NullHandler())